Big ugly commit of all the changes made on site at LASCOT
These are all the changes I made as I worked against the clock to get the
code working in time for the talk.  This includes changing the palette
colours on the IMPs and adding in extra calls to transmit to work around
UDP messages getting lost.

The next thing to do is to switch from UDP to TCP to get rid of this
issue.

There is also routing responsibility in the message class that really
belongs in the node.  Messages don't know how to route themselves, nodes do.

Also added the network class to bundle a bunch of nodes together and get
them all to process their messages so we can step through a scripted demo.

Finally, on the day before the talk I had the idea of using a couple of the
nodes as hosts rather than IMPS.  These allowed me to show a message leaving
the host at UCLA, going via the UCLA and the SRI IMP and ending up on the SRI
host.  On the hosts messages are not rotated through the buffers as they
are on the IMPs.  Instead they move tetris style through the buffer and are
only removed when all three message slots are full and a new message comes in.

I need to have another look at the routing code as something funny was
happening to messages going back the other way when I changed the weighting
of the UCLA -> SRI (And so the SRI -> UCLA) link to 10.  It was failing with
a KeyError.  Didn't have time to troubleshoot before the talk.